2010-06-29  Lin Ma  <Lin.Ma@Sun.COM>
* capplet/nwam_wireless_dialog.c
(nwam_wireless_dialog_set_key_index),
(nwam_wireless_dialog_get_key_index):
* capplet/nwam_wireless_dialog.h: Fixed 16390 [osol_142]
nwam-manager cores when joining a unlisted wireless network with
WEP security

git-svn-id: file:///export/mirror/svn/svn.opensolaris.org/svn/jds/nwam-manager/trunk@668 6b7d1fb9-75dd-c4d7-8f26-8a8fef618664

+++ capplet/nwam_wireless_dialog.h
@@ -97,9 +97,9 @@ extern  void                    nwam_wireless_dialog_set_key (NwamWirelessDialog
 extern  gchar*                  nwam_wireless_dialog_get_key (NwamWirelessDialog *wireless_dialog );
 
 extern  void                    nwam_wireless_dialog_set_key_index (NwamWirelessDialog  *wireless_dialog,
-                                                              guint key_index );
+                                                              guint64 key_index );
 
-extern  guint                   nwam_wireless_dialog_get_key_index (NwamWirelessDialog *wireless_dialog );
+extern  guint64                   nwam_wireless_dialog_get_key_index (NwamWirelessDialog *wireless_dialog );
 
 extern  void                    nwam_wireless_dialog_set_bssid_list (NwamWirelessDialog  *wireless_dialog,
                                                                      GList               *bssid_list );
